pyproject.toml
[project]
name = "longform-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "django>=5.2.10",
    "google-genai>=1.59.0",
    "google-generativeai>=0.8.6",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
